adversarial_icl/requirement.txt
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.64.0
pyyaml>=6.0
transformers>=4.30.0
torchvision>=0.15.0